==> Scripts/4-analyze-output.R <==
#################
# 4 - Analyze model output
#
# Author: Casey Youngflesh
#################



# Clear environment -------------------------------------------------------

rm(list = ls())



# Load packages -----------------------------------------------------------

if('pacman' %in% rownames(installed.packages()) == FALSE)
{
  install.packages('pacman')
}

pacman::p_load(MCMCvis, boot, dplyr)




# Load data -------------------------------------------------------

#phi = survival prob
#p = detection prob

#NAME <- 'PW_60k_2019-04-12_FULL_beta[j,k]' #nu_p[i,j,k] + beta_p[j,k]
NAME <- 'PW_100k_2019-07-10_nu_p[i,j,k]_beta[j,k]'
OUTPUT <- '~/Google_Drive/R/penguin_watch_model/Results/OUTPUT-2019-07-10'

setwd(paste0('~/Google_Drive/R/penguin_watch_model/Results/', NAME))

fit <- readRDS(paste0(NAME, '.rds'))
data <- readRDS('jagsData.rds')

setwd('~/Google_Drive/R/penguin_watch_model/Data/PW_data/')
PW_data <- read.csv('PW_data_2019-07-10.csv', stringsAsFactors = FALSE)



# summaries ---------------------------------------------------------------

sm <- MCMCvis::MCMCsummary(fit, excl = c('z_out', 'p_out'))
max(sm[,'Rhat'])
hist(sm[,'Rhat'])
min(sm[,'n.eff'])
hist(sm[,'n.eff'])



# trace plots -------------------------------------------------------------

#create dir for figs if doesn't exist
ifelse(!dir.exists(OUTPUT), 
       dir.create(OUTPUT), 
       FALSE)

setwd(OUTPUT)

# MCMCvis::MCMCtrace(fit, params = 'z_out', Rhat = TRUE, n.eff = TRUE,
#                    filename = 'z_out_trace.pdf')
MCMCvis::MCMCtrace(fit, params = 'nu_p', Rhat = TRUE, n.eff = TRUE,
                   ind = TRUE, filename = 'nu_p_trace.pdf')
MCMCvis::MCMCtrace(fit, params = 'beta_p', Rhat = TRUE, n.eff = TRUE,
                   ind = TRUE, filename = 'beta_p_trace.pdf')
MCMCvis::MCMCtrace(fit, params = 'mu_phi', Rhat = TRUE, n.eff = TRUE,
                   ind = TRUE, filename = 'mu_phi_trace.pdf')


# plot BS ------------------------------------------

#mu_phi
mu_phi <- MCMCvis::MCMCpstr(fit, params = 'mu_phi')[[1]]
mu_phi_LCI <- MCMCvis::MCMCpstr(fit, params = 'mu_phi',
                                func = function(x) quantile(x, probs = c(0.025)))[[1]]
mu_phi_UCI <- MCMCvis::MCMCpstr(fit, params = 'mu_phi',
                                func = function(x) quantile(x, probs = c(0.975)))[[1]]

#mu_phi_bs
mu_phi_bs <- MCMCvis::MCMCpstr(fit, params = 'mu_phi_bs')[[1]]
mu_phi_bs_LCI <- MCMCvis::MCMCpstr(fit, params = 'mu_phi_bs',
                                func = function(x) quantile(x, probs = c(0.025)))[[1]]
mu_phi_bs_UCI <- MCMCvis::MCMCpstr(fit, params = 'mu_phi_bs',
                                func = function(x) quantile(x, probs = c(0.975)))[[1]]


# setwd('~/Google_Drive/R/penguin_watch_model/Data/')
# SLL <- read.csv('site_ll.csv')

# setwd('Krill_data/CCAMLR/Processed_CCAMLR/')
# krill <- read.csv('CCAMLR_krill_entire_season.csv')
# 
# setwd('../../../SIC_data/Processed')
# sea_ice <- read.csv('SIC_150_W.csv')
# sea_ice2 <- sea_ice[,c(1,2,7)]

#USE d_mrg in data object
yrs_rng <- range(data$yrs_array, na.rm = TRUE)

site_vec <- c()
year_vec <- c()
for (i in 1:length(data$unsites))
{
  #i <- 1
  tv <- rep(data$unsites[i], sum(!is.na(data$yrs_array[,i])))
  yv <- data$yrs_array[which(!is.na(data$yrs_array[,i])),i]
  
  site_vec <- c(site_vec, tv)
  year_vec <- c(year_vec, yv)
}

mrg2 <- data.frame(SITE = site_vec,
                  YEAR = year_vec,
                  mn_mu_phi = as.vector(mu_phi_bs)[!is.na(as.vector(mu_phi_bs))],
                  LCI_mu_phi = as.vector(mu_phi_bs_LCI)[!is.na(as.vector(mu_phi_bs))],
                  UCI_mu_phi = as.vector(mu_phi_bs_UCI)[!is.na(as.vector(mu_phi_bs))])


library(ggplot2)
p <- ggplot(mrg2, aes(YEAR, mn_mu_phi, color = SITE)) + 
  geom_point(size = 3, position = position_dodge(width = 0.7)) +
  geom_errorbar(data = mrg2,
                aes(ymin = LCI_mu_phi, ymax = UCI_mu_phi,
                    color = SITE), width = 1.1,
                position = position_dodge(width = 0.7)) +
  theme_bw() +
  ylim(c(0, 2)) +
  ylab('Breeding Success') +
  xlab('Year')

ggsave('site_bs.pdf', p)




# plot regression fit SIC/KRILL -----------------------------------------------------

# plot(mrg5$T_KRILL, mrg5$mn_mu_phi)
# plot(mrg5$W_MN, mrg5$mn_mu_phi)
# plot(mrg5$latitude, mrg5$mn_mu_phi)
# 
# 
# alpha_ch <- MCMCvis::MCMCchains(fit, params = 'alpha_theta')[,1]
# pi_ch <- MCMCvis::MCMCchains(fit, params = 'pi_theta')[,1]
# rho_ch <- MCMCvis::MCMCchains(fit, params = 'rho_theta')[,1]
# 
# sim_KRILL <- seq(min(data$KRILL, na.rm = TRUE)-1, 
#                  max(data$KRILL, na.rm = TRUE)+1, length = 100)
# sim_SIC <- seq(min(data$SIC, na.rm = TRUE)-1, 
#                max(data$SIC, na.rm = TRUE)+1, length = 100)
# 
# mf_KRILL <- matrix(nrow = length(alpha_ch), ncol = 100)
# mf_SIC <- matrix(nrow = length(alpha_ch), ncol = 100)
# for (i in 1:length(sim_KRILL))
# {
#   mf_KRILL[,i] <- alpha_ch + rho_ch * sim_KRILL[i] #+ pi_ch * mean(sim_SIC) #mean of cov is 0
#   mf_SIC[,i] <- alpha_ch + pi_ch * sim_SIC[i] #+ rho_ch * mean(sim_KRILL) #mean of cov is 0
# }
# 
# med_mf_KRILL <- apply(mf_KRILL, 2, median)
# LCI_mf_KRILL <- apply(mf_KRILL, 2, function(x) quantile(x, probs = 0.025))
# UCI_mf_KRILL <- apply(mf_KRILL, 2, function(x) quantile(x, probs = 0.975))
# 
# FIT_PLOT <- data.frame(MN_FIT = med_mf_KRILL, 
#                        MN_X = sim_KRILL,
#                        LCI_FIT = LCI_mf_KRILL,
#                        UCI_FIT = UCI_mf_KRILL)
# 
# DATA_PLOT <- data.frame(MN_phi = as.vector(mu_phi), 
#                        LCI_phi = as.vector(mu_phi_LCI),
#                        UCI_phi = as.vector(mu_phi_UCI),
#                        KRILL = as.vector(data$KRILL),
#                        SIC = as.vector(data$SIC))
# 
# ggplot(data = DATA_PLOT, aes(KRILL, MN_phi), color = 'black', alpha = 0.6) +
#   geom_ribbon(data = FIT_PLOT, 
#               aes(x = MN_X, ymin = LCI_FIT, ymax = UCI_FIT),
#               fill = 'grey', alpha = 0.7,
#               inherit.aes = FALSE) +
#   geom_line(data = FIT_PLOT, aes(MN_X, MN_FIT), color = 'red',
#             alpha = 0.9,
#             inherit.aes = FALSE,
#             size = 1.4) +
#   geom_errorbar(data = DATA_PLOT, 
#                 aes(ymin = LCI_phi, ymax = UCI_phi), width = 0.3,
#                 color = 'black', alpha = 0.2) +
#   geom_point(data = DATA_PLOT, aes(KRILL, MN_phi), color = 'black',
#              inherit.aes = FALSE, size = 3, alpha = 0.7) +
#   theme_bw() +
#   xlab('Krill catch') +
#   ylab('survival rate') +
#   theme(
#     axis.text = element_text(size = 16),
#     axis.title = element_text(size = 18),
#     axis.title.y = element_text(margin = margin(t = 0, r = 15, b = 0, l = 0)),
#     axis.title.x = element_text(margin = margin(t = 15, r = 15, b = 0, l = 0)),
#     axis.ticks.length= unit(0.2, 'cm')) #length of axis tick
# 
# 
# med_mf_SIC <- apply(mf_SIC, 2, median)
# LCI_mf_SIC <- apply(mf_SIC, 2, function(x) quantile(x, probs = 0.025))
# UCI_mf_SIC <- apply(mf_SIC, 2, function(x) quantile(x, probs = 0.975))
# 
# FIT_PLOT <- data.frame(MN_FIT = med_mf_SIC, 
#                        MN_X = sim_SIC,
#                        LCI_FIT = LCI_mf_SIC,
#                        UCI_FIT = UCI_mf_SIC)
# 
# ggplot(data = DATA_PLOT, aes(SIC, MN_phi), color = 'black', alpha = 0.6) +
#   geom_ribbon(data = FIT_PLOT, 
#               aes(x = MN_X, ymin = LCI_FIT, ymax = UCI_FIT),
#               fill = 'grey', alpha = 0.7,
#               inherit.aes = FALSE) +
#   geom_line(data = FIT_PLOT, aes(MN_X, MN_FIT), color = 'red',
#             alpha = 0.9,
#             inherit.aes = FALSE,
#             size = 1.4) +
#   geom_errorbar(data = DATA_PLOT, 
#                 aes(ymin = LCI_phi, ymax = UCI_phi), width = 0.3,
#                 color = 'black', alpha = 0.2) +
#   geom_point(data = DATA_PLOT, aes(SIC, MN_phi), color = 'black',
#              inherit.aes = FALSE, size = 3, alpha = 0.7) +
#   theme_bw() +
#   xlab('SIC') +
#   ylab('survival rate') +
#   theme(
#     axis.text = element_text(size = 16),
#     axis.title = element_text(size = 18),
#     axis.title.y = element_text(margin = margin(t = 0, r = 15, b = 0, l = 0)),
#     axis.title.x = element_text(margin = margin(t = 15, r = 15, b = 0, l = 0)),
#     axis.ticks.length= unit(0.2, 'cm')) #length of axis tick



# plotly 3D plot ----------------------------------------------------------

# #from here: https://stackoverflow.com/questions/36049595/mixing-surface-and-scatterplot-in-a-single-3d-plot
# 
# mf_KR_SIC <- matrix(nrow =  length(sim_KRILL), ncol = length(sim_SIC))
# d_pts <- data.frame(MP = as.vector(mu_phi),
#                     KR = as.vector(data$KRILL),
#                     S = as.vector(data$SIC))
# for (i in 1:length(sim_KRILL))
# {
#   for (j in 1:length(sim_SIC))
#   {
#     mf_KR_SIC[i,j] <- mean(alpha_ch) + 
#       mean(rho_ch) * sim_KRILL[i] + 
#       mean(pi_ch) * sim_SIC[j]
#   }
# }
# 
# 
# library(plotly)
# plot_ly(x = sim_KRILL,
#         y = sim_SIC,
#         z = mf_KR_SIC, type = "surface") %>% 
#   add_trace(data = d_pts, x = d_pts$KR, y = d_pts$S, z = d_pts$MP, 
#             mode = "markers", type = "scatter3d", 
#             marker = list(size = 5, color = "red", symbol = 104))


# precip events -----------------------------------------------------------

#aggregate precip events to day (sum rain and max snow)

setwd('~/Google_Drive/R/penguin_watch_model/Data/precip_data')

fls <- list.files()

precip_df <- data.frame()
for (i in 1:length(data$unsites))
{
  #i <- 1
  fls2 <- fls[grep(data$unsites[i], fls)]
  
  years <- as.numeric(substr(fls2, start = 6, stop = 9))
  for (j in 1:length(years))
  {
    #j <- 1
    
    #read in csv for that site/year
    tt <- read.csv(fls2[grep(years[j], fls2)])
    
    #remove time and transform to date
    dates <- as.Date(sapply(strsplit(as.character(tt$datetime), 
                                    split = ' '), '[', 1), 
                    format = '%Y:%m:%d')
    
    udates <- unique(dates)
    for (k in 1:length(udates))
    {
      #k <- 1
      t2 <- tt[which(dates == udates[k]),]
      
      #max snow score
      t_snow <- t2$score[grep('S', t2$score)]
      if (length(t_snow) > 0)
      {
        m_snow <- max(as.numeric(substr(t_snow, start = 2, stop = 2)))
      } else {
        m_snow <- 0
      }
      
      #rain
      t_rain <- t2$score[grep('R', t2$score)]
      if (length(t_rain) > 0)
      {
        s_rain <- sum(as.numeric(substr(t_rain, start = 2, stop = 2)))
      } else {
        s_rain <- 0
      }
      
      t_precip <- data.frame(site = data$unsites[i],
                             season_year = years[j],
                             date = udates[k],
                             m_snow,
                             s_rain)
      
      precip_df <- rbind(precip_df, t_precip)
    }
  }
}




# time varying plots -----------------------------------------------------

z_out_mn <- MCMCvis::MCMCpstr(fit, params = 'z_out', func = mean)[[1]]
z_out_sd <- MCMCvis::MCMCpstr(fit, params = 'z_out', func = sd)[[1]]
z_out_LCI <- z_out_mn - z_out_sd
z_out_UCI <- z_out_mn + z_out_sd

p_out_mn <- MCMCvis::MCMCpstr(fit, params = 'p_out', func = mean)[[1]]
p_out_sd <- MCMCvis::MCMCpstr(fit, params = 'p_out', func = sd)[[1]]
p_out_LCI <- p_out_mn - p_out_sd
p_out_UCI <- p_out_mn + p_out_sd

#plot snow events in purple (2 or greater) and rain events in orange (2 or greater)
#width of vertical lines indicate severity of precipitation event

#create dir for figs if doesn't exist
ifelse(!dir.exists(paste0(OUTPUT, '/time_plots')), 
       dir.create(paste0(OUTPUT, '/time_plots')), 
       FALSE)

setwd(paste0(OUTPUT, '/time_plots'))

for (i in 1:dim(data$date_array)[3])
{
  #i <- 1
  #remove years that don't have data (wasn't done in model script)
  t_date <- data$date_array[,,i]
  to.rm.dt <- which(is.na(t_date[1,]))
  if (length(to.rm.dt) > 0)
  {
    t_date2 <- t_date[,-to.rm.dt]
  }
  for (j in 1:dim(z_out_mn)[2])
  {
    #j <- 2
    
    if (sum(!is.na(z_out_mn[,j,i])) > 0)
    {
      #if t_date2 doesn't have dims (all NA cols were removed)
      if (is.null(dim(t_date2)))
      {
        dates <- as.Date(t_date2, origin = '1970-01-01')
      } else {
        dates <- as.Date(t_date2[,j], origin = '1970-01-01')
      }
      
      #filter for daily precip events for site/year
      t_precip <- dplyr::filter(precip_df, 
                                site == data$unsites[i],
                                season_year == data$yrs_array[j,i],
                                m_snow >= 2 | s_rain >= 2,
                                date < max(dates),
                                date > min(dates))
      t_precip$ts <- as.numeric(t_precip[,'date'] - min(dates))
      
      
      #keep every 5th date value
      n_dates <- dates[seq(1, 60, by = 5)]
      #breaks for x axis on plot
      n_breaks <- seq(1, 60, by = 5)
      #min number of chicks (from counts in images and known counts later)
      counts <- data$c_array[,j,i] 
      #remove 0 vals
      to.na <- which(counts == 0)
      if (length(to.na) > 0)
      {
        counts[to.na] <- NA
      }
      
      #NEEDS TO BE CHANGED AFTER NEXT MODEL RUN TO REFLECT DIFFERENT STARTING POINTS
      #dotted line from start of season to first chick
      dt_seq <- seq(z_out_mn[1,j,i], z_out_mn[31,j,i], length = 31)
      
      SITE <- data$unsites[i]
      YEAR <- data$yrs_array[j,i]
      min_z_out <- min(z_out_LCI[,j,i])
      rng_z_out <- max(z_out_UCI[,j,i]) - min_z_out
      rng_dt_seq <- (range(dt_seq)[2])
      PLT_DF <- data.frame(time = 1:60,  
                           z_out_mn = c(rep(NA, 30), z_out_mn[31:60,j,i]),
                           z_out_LCI = c(rep(NA, 30), z_out_LCI[31:60,j,i]),
                           z_out_UCI = c(rep(NA, 30), z_out_UCI[31:60,j,i]),
                           z_out_dot = c(dt_seq, rep(NA, 29)),
                           p_out_mn = p_out_mn[,j,i], 
                           p_out_LCI = p_out_LCI[,j,i],
                           p_out_UCI = p_out_UCI[,j,i],
                           # p_out_mn_sc = p_out_mn[,j,i] * rng_z_out + min_z_out,
                           # p_out_LCI_sc = p_out_LCI[,j,i] * rng_z_out + min_z_out,
                           # p_out_UCI_sc = p_out_UCI[,j,i] * rng_z_out + min_z_out,
                           p_out_mn_sc = p_out_mn[,j,i] * rng_dt_seq,
                           p_out_LCI_sc = p_out_LCI[,j,i] * rng_dt_seq,
                           p_out_UCI_sc = p_out_UCI[,j,i] * rng_dt_seq,
                           count = counts)
      
      p <- ggplot(PLT_DF, aes(x = time)) + 
        geom_ribbon(aes(ymin = z_out_LCI, ymax = z_out_UCI),
                  fill = 'blue', alpha = 0.2) +
        geom_line(aes(y = z_out_mn), col = 'blue') +
        geom_line(aes(y = z_out_dot), col = 'blue', linetype = 2) +
        geom_ribbon(aes(ymin = p_out_LCI_sc, ymax = p_out_UCI_sc),
                    fill = 'red', alpha = 0.2) +
        geom_line(aes(y = p_out_mn_sc),
                  col = 'red') +
        #precipitation
        geom_vline(xintercept = t_precip$ts, 
                       color = 'purple', size = (t_precip$m_snow/3),
                   alpha = 0.5) +
        geom_vline(xintercept = t_precip$ts, 
                   color = 'orange', size = (t_precip$s_rain/3),
                   alpha = 0.5) +
        # geom_line(aes(y = count),
        #           col = 'green') +
        theme_bw() +
        scale_y_continuous(limits = c(-1, (max(dt_seq) + 1)),
                           sec.axis = sec_axis(~(.)/max(dt_seq),
                                               name = 'Detection probability')) +
        # scale_y_continuous(sec.axis = sec_axis(~(.-min_z_out)/rng_z_out,
        #                                        name = 'Detection probability')) +
        scale_x_continuous(labels = n_dates, breaks = n_breaks) +
        ylab('Number of chicks') +
        xlab('') +
        ggtitle(paste0(SITE, ' - ', YEAR)) + 
        theme(axis.text.x = element_text(angle = 90, hjust = 1))
      
      #print(p)
      ggsave(p, filename = paste0(SITE, '-', YEAR, '-2019-07-10.pdf'))
    }
  }
}




# analyze -----------------------------------------------------------------

#to check where there might be problems - reference plots created above and look at time series
data$unsites
data$yrs_array

site <- 12
year <- 2
tdat <- data$y[-c(1:30), , year, site]
n_id <- which(apply(tdat, 2, function(x) sum(!is.na(x))) > 0)

tdat[, n_id]



# compare death rates at different period of season -----------------------

mort <- data.frame()
for (i in 1:length(data$unsites))
{
  #i <- 1
  for (j in 1:length(data$yrs_array[,i]))
  {
    #j <- 5
    if (!is.na(data$yrs_array[j,i]))
    {
      #lay to hatch
      lh <- round((z_out_mn[1,j,i] - z_out_mn[31,j,i]) / 31, 3)
      
      #young hatch to older hatch
      yo <- round((z_out_mn[31,j,i] - z_out_mn[45,j,i]) / 15, 3)
      
      #older hatch to creche
      oc <- round((z_out_mn[45,j,i] - z_out_mn[60,j,i]) / 15, 3)
      
      t_m <- data.frame(site = data$unsites[i], 
                        season_year = data$yrs_array[j,i],
                        lh, yo, oc)
      
      mort <- rbind(mort, t_m)
    }
  }
}

num <- cbind(rep(1, NROW(mort)), rep(2, NROW(mort)), rep(3, NROW(mort)))
mval <- mort[,c('lh', 'yo', 'oc')]




setwd(OUTPUT)

pdf('death_rates_season.pdf', width = 5, height = 5)
plot(density(mval[,1]), xlim = c(0, 0.6), ylim = c(0,5), 
     col = rgb(1,0,0,0.3), lwd = 3, 
     main = 'Chick death rates over nesting season', 
     xlab = 'Chick deaths per day')
lines(density(mval[,2]), col = rgb(0,1,0,0.3), lwd = 3)
lines(density(mval[,3]), col = rgb(0,0,1,0.3), lwd = 3)
legend('topright',
       legend = c('lay -> hatch', 'hatch -> young chick', 
                  'young chick -> creche'),
       col = c(rgb(1,0,0,0.3), rgb(0,1,0,0.3), rgb(0,0,1,0.3)),
       lwd = c(2,2,2,2), cex = 1)
dev.off()

#friedman test to check for differences in means (repeated measures - does not assume normality)
friedman.test(as.matrix(mval))
#post-hoc test
require(PMCMR)
posthoc.friedman.nemenyi.test(as.matrix(mval))




# merge data ----------------------------------------------------------

p2 <- data.frame()
for (i in 1:length(data$unsites))
{
  #i <- 1
  tsite <- dplyr::filter(precip_df, site == data$unsites[i])
  
  u_year <- unique(tsite$season_year)
  for (j in 1:length(u_year))
  {
    #j <- 3
    tyear <- dplyr::filter(tsite, season_year == u_year[j])
    tsnow <- length(which(tyear$m_snow >= 2))
    train <- length(which(tyear$s_rain >= 2))
    
    tt <- data.frame(SITE = data$unsites[i], 
                     YEAR = u_year[j],
                     tsnow,
                     train)
    p2 <- rbind(p2, tt)
  }
}

#merge lat/ln with precip, BS, and creche date
cll <- unique(PW_data[,c('site', 'col_lat', 'col_lon')])
mrg3 <- dplyr::left_join(mrg2, cll, by = c('SITE' = 'site'))
mrg4 <- dplyr::left_join(mrg3, p2, by = c('SITE', 'YEAR'))

mrg5 <- dplyr::left_join(mrg4, data$d_mrg, by = c('SITE' = 'site', 'YEAR' = 'season_year'))


#Merge with Hinke et al. 2017 (MEE) data
hinke_2017 <- data.frame(SITE = c('CAPE', 'CIER', 'COPA', 'GALI', 'LION', 'PETE'), 
                         YEAR = rep(2017, 6),
                         mn_mu_phi = c(1.63, 1.47, 1.53, 1.46, 1.26, 1.51),
                         LCI_mu_phi = rep(NA, 6),
                         UCI_mu_phi = rep(NA, 6),
                         col_lat = c(-62.46, -64.143, -62.175, 
                                     -65.244, -62.135, -65.177),
                         col_lon = c(-60.789, -60.984, -58.456, 
                                     -64.247, -58.126, -64.189),
                         tsnow = rep(NA, 6),
                         train = rep(NA, 6),
                         j = rep(NA, 6),
                         k = rep(NA, 6),
                         num_nests = c(8, 15, 58, 28, 19, 37),
                         chick_date = rep(NA, 6),
                         creche_date = c('2017-01-18', '2017-01-20', '2016-12-22',
                                         '2017-02-04', '2016-12-20', '2017-02-08'),
                         days = rep(NA, 6))

mrg6 <- rbind(mrg5, hinke_2017)

mrg6$SOURCE <- c(rep('PW', length(which(!is.na(mrg6$train)))), 
                 rep('Hinke', length(which(is.na(mrg6$train)))))


#merge with krill data
setwd('~/Google_Drive/R/penguin_watch_model/Data/Krill_data/CCAMLR/Processed_CCAMLR/')

#krill catch from entire year
krill <- read.csv('CCAMLR_krill_entire_season.csv')
colnames(krill)[3] <- 'YR_KRILL'
mrg7 <- dplyr::left_join(mrg6, krill, by = c('SITE', 'YEAR'))

#krill catch from just breeding season
krill2 <- read.csv('CCAMLR_krill_breeding_season.csv')
colnames(krill2)[3] <- 'BR_KRILL'
mrg8 <- dplyr::left_join(mrg7, krill2, by = c('SITE', 'YEAR'))


# covariates/plots --------------------------------------------------------------

#compare BS to total precip - no relationship
plot(mrg8$tsnow, mrg8$mn_mu_phi)
plot(mrg8$train, mrg8$mn_mu_phi)



#compare BS to phenology
#convert creche date to julian day
jd_1 <- as.numeric(strftime(mrg8$creche_date, format = "%j"))
idx <- which(jd_1 > 300)
jd_2 <- jd_1
jd_2[idx] <- jd_1[idx] - 300
jd_2[-idx] <- jd_1[-idx] + 65

mrg8$c_jd <- jd_2

plot(mrg8$c_jd, mrg8$mn_mu_phi)
plot(mrg8$col_lat, mrg8$c_jd)


#compare BS to lat and substrate type
# to.rm <- which(mrg4$col_lat > -60)
# mrg5 <- mrg4[-to.rm,]
# plot(mrg5$col_lat, mrg5$mn_mu_phi)
# f1 <- lm(mn_mu_phi ~ col_lat, data = mrg5)
# abline(f1, col = 'red')

#compare BS to krill
library(quantreg)
plot(mrg8$YR_KRILL, mrg8$mn_mu_phi)



# BS map ------------------------------------------------------------------

#map of breeding success
require(raster)
setwd('~/Google_Drive/R/penguin_watch_model/Data/peninsula/')
AP <- rgdal::readOGR('GADM_peninsula.shp')

mrg6_agg <- aggregate(mn_mu_phi ~ SITE + col_lat + col_lon + SOURCE, data = mrg6, mean)

setwd(OUTPUT)

#all sites
pdf('BS_map.pdf')
#AP shp file
ggplot(data = AP, aes(long, lat, group = group)) +
  geom_polygon(fill = 'grey') + 
  geom_path(data = AP, aes(long, lat, group = group), 
            inherit.aes = FALSE,
            color = 'black') +
  #all coord
  coord_map(xlim = c(-68, -33),
              ylim = c(-67, -51)) +
  theme_void() +
  #theme_bw()
  #BS
  geom_point(data = mrg6_agg,
             inherit.aes = FALSE,
             size = 8,
             alpha = 0.9,
             aes(col_lon, col_lat, color = mn_mu_phi)) +
  scale_color_gradient('Chicks per pair',
                       limits = c(min(mrg6$mn_mu_phi),
                                  max(mrg6$mn_mu_phi)),
                       low = '#2c7fb8',
                       high = '#edf8b1') +
  # #point outlines
  # geom_point(data = mrg6_agg,
  #            inherit.aes = FALSE,
  #            size = 8,
  #            shape = 21,
  #            alpha = 0.8,
  #            stroke = 1,
  #            color = 'black',
  #            aes(col_lon, col_lat)) +
  #point outlines - PW
  geom_point(data = mrg6_agg[which(mrg6_agg$SOURCE == 'PW'),],
             inherit.aes = FALSE,
             size = 8,
             shape = 21,
             alpha = 0.8,
             stroke = 1,
             color = 'black',
             aes(col_lon, col_lat)) +
  #point outlines - Hinke
  geom_point(data = mrg6_agg[which(mrg6_agg$SOURCE == 'Hinke'),],
             inherit.aes = FALSE,
             size = 8,
             shape = 21,
             alpha = 0.8,
             stroke = 1,
             color = 'red',
             aes(col_lon, col_lat))
  #theme(legend.position='none') +
dev.off()


#just AP sites
pdf('BS_map_AP.pdf')
#AP shp file
ggplot(data = AP, aes(long, lat, group = group)) +
  geom_polygon(fill = 'grey') + 
  geom_path(data = AP, aes(long, lat, group = group), 
            inherit.aes = FALSE,
            color = 'black') +
  #AP coord
  coord_map(xlim = c(-68, -53),
              ylim = c(-66.5, -61)) +
  # #all coord
  # coord_map(xlim = c(-68, -33),
  #           ylim = c(-67, -51)) +
  theme_void() +
  #theme_bw()
  #BS
  geom_point(data = mrg6_agg,
             inherit.aes = FALSE,
             size = 8,
             alpha = 0.9,
             aes(col_lon, col_lat, color = mn_mu_phi)) +
  scale_color_gradient('Chicks per pair',
                       limits = c(min(mrg6$mn_mu_phi),
                                  max(mrg6$mn_mu_phi)),
                       low = '#2c7fb8',
                       high = '#edf8b1') +
  # #point outlines
  # geom_point(data = mrg6_agg,
  #            inherit.aes = FALSE,
  #            size = 8,
  #            shape = 21,
  #            alpha = 0.8,
  #            stroke = 1,
  #            color = 'black',
  #            aes(col_lon, col_lat)) +
  #point outlines - PW
  geom_point(data = mrg6_agg[which(mrg6_agg$SOURCE == 'PW'),],
             inherit.aes = FALSE,
             size = 8,
             shape = 21,
             alpha = 0.8,
             stroke = 1,
             color = 'black',
             aes(col_lon, col_lat)) +
  #point outlines - Hinke
  geom_point(data = mrg6_agg[which(mrg6_agg$SOURCE == 'Hinke'),],
             inherit.aes = FALSE,
             size = 8,
             shape = 21,
             alpha = 0.8,
             stroke = 1,
             color = 'red',
             aes(col_lon, col_lat))
#theme(legend.position='none') +
dev.off()







# PPO ---------------------------------------------------------------------


tf <- function(PR)
{
  hist(inv.logit(PR))
}

# mu_p ~ dnorm(2, 0.1)
PR <- rnorm(15000, 0, 1/sqrt(0.1))
tf(PR)
MCMCtrace(fit, 
          params = 'mu_p',
          ind = TRUE, 
          priors = PR,
          pdf = FALSE,
          post_zm = FALSE)


# mu_beta_p ~ dnorm(0.1, 10) T(0, 0.5)
PR_p <- rnorm(15000, 0.1, 1/sqrt(10))
PR <- PR_p[which(PR_p > 0 & PR_p < 0.5)]
tf(PR)
MCMCtrace(fit, 
          params = 'mu_beta_p',
          ind = TRUE, 
          priors = PR,
          pdf = FALSE,
          post_zm = FALSE)

# sigma_beta_p ~ dunif(0, 2)
PR <- runif(15000, 0, 2)
MCMCtrace(fit, 
          params = 'sigma_beta_p',
          ind = TRUE, 
          priors = PR,
          pdf = FALSE,
          post_zm = FALSE)

# theta_phi ~ dnorm(4, 0.25)
PR <- rnorm(15000, 4, 1/sqrt(0.25))
tf(PR)
MCMCtrace(fit, 
          params = 'theta_phi',
          ind = TRUE, 
          priors = PR,
          pdf = FALSE,
          post_zm = FALSE)

# sigma_mu_phi ~ dunif(0, 3)
PR <- runif(15000, 0, 3)
MCMCtrace(fit, 
          params = 'sigma_mu_phi',
          ind = TRUE, 
          priors = PR,
          pdf = FALSE,
          post_zm = FALSE)

# sigma_nu_p ~ dunif(0, 3)
PR <- runif(15000, 0, 3)
MCMCtrace(fit, 
          params = 'sigma_nu_p',
          ind = TRUE, 
          priors = PR,
          pdf = FALSE,
          post_zm = FALSE)
